Add thread specific clock
Standard cpp clocks count the absolute number of ticks that passed, even
if the thread was sleeping during that time.
This might create significant deviations and corrupt the benchmarking
data in certain applications.
This is a wrapper for the POSIX thead-specific clock, that complies with
`std::chrono`-s standard, allowing us to use it as THE clock for our
Benchmarking infrastructure.